requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-Testing==0.4.2
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==2.0.2
Werkzeug==0.10.4
nose==1.3.7
coverage==3.7.1

#ipython==3.2.1
#tornado==4.2.1
#pyzmq==14.7.0
#backports.ssl-match-hostname==3.4.0.2
#certifi==14.5.14
#jsonschema==2.4.0
#ptyprocess==0.4
#mistune==0.7
#terminado==0.5

